requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
cloudinary==1.36.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
pydantic==1.10.12
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
realesrgan==0.3.0
psutil==5.9.6
